+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v2.1.1] - 2022-06-15
+
 - Add padding to item information UI
 - Add log message if there is an error monitoring a directory
 
@@ -30,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add appstream info file
 - Various style changes
 
-[unreleased]: https://github.com/EbonJaeger/budgie-trash-applet/compare/v2.1.0...master
+[unreleased]: https://github.com/EbonJaeger/budgie-trash-applet/compare/v2.1.1...master
+[v2.1.1]: https://github.com/EbonJaeger/budgie-trash-applet/compare/v2.1.0...v2.1.1
 [v2.1.0]: https://github.com/EbonJaeger/budgie-trash-applet/compare/v2.0.0...v2.1.0
 [v2.0.0]: https://github.com/EbonJaeger/budgie-trash-applet/compare/v1.2.0...v2.0.0